Fixed Queue Problem with Ant movement    -Had to replace the move with moveAlongPath    -A* needs integer starts and ends, if the ant was inbetween a    location it had a float.
TODO
    - Still an odd movement that occurs when ant moves right and then
      redirects left (or vice-versa).  For instance, if you click to the
      right of the ant, and then before it gets to the destination you
      click below it and to the left the ant kinda skips a tile or
      something.

+++ src/Ant.py
@@ -321,8 +321,8 @@ class Ant():
 
     # Find a path using A* Manhattan
     def findPath(self):
-        start = [self.pos[0] / 32, self.pos[1] / 32]
-        end = [self.newPos[0] / 32, self.newPos[1] / 32]
+        start = [int(self.pos[0] / 32), int(self.pos[1] / 32)]
+        end = [int(self.newPos[0] / 32), int(self.newPos[1] / 32)]
         
         # Start and end are the same tile, dont need to move.
         if start == end:
